Consolidate empty program and data checking.
Fixes #111, fixes #113.
