+++ src/activateDaffodilDebug.ts
@@ -7,8 +7,6 @@
 import * as vscode from 'vscode'
 import * as path from 'path'
 import * as hexView from './hexview/hexView'
-import * as os from 'os'
-import * as child_process from 'child_process'
 import {
   WorkspaceFolder,
   DebugConfiguration,
@@ -19,25 +17,7 @@ import { DaffodilDebugSession } from './daffodilDebug'
 import { getDebugger, getDataFileFromFolder } from './daffodilDebugger'
 import { FileAccessor } from './daffodilRuntime'
 import * as fs from 'fs'
-import XDGAppPaths from 'xdg-app-paths'
-const xdgAppPaths = XDGAppPaths({ name: 'dapodil' })
 import * as infoset from './infoset'
-import { deactivate } from './extension'
-
-// Function for stopping debuggin
-function stopDebugging() {
-  vscode.debug.stopDebugging()
-  deactivate()
-  vscode.window.activeTerminal?.processId.then((id) => {
-    if (id) {
-      if (os.platform() === 'win32') {
-        child_process.exec(`taskkill /F /PID ${id}`)
-      } else {
-        child_process.exec(`kill -9 ${id}`)
-      }
-    }
-  })
-}
 
 // Function for setting up the commands for Run and Debug file
 function createDebugRunFileConfigs(resource: vscode.Uri, runOrDebug: String) {
@@ -133,25 +113,7 @@ export function activateDaffodilDebug(
 
             return ''
           })
-
-        // Create file that holds path to program file used
-        await fs.writeFile(
-          `${xdgAppPaths.data()}/.programFile`,
-          programFile,
-          function (err) {
-            if (err) {
-              vscode.window.showInformationMessage(
-                `error code: ${err.code} - ${err.message}`
-              )
-            }
-          }
-        )
-
-        // If program file not selected stop launch
-        if (programFile === '') {
-          stopDebugging()
-        }
-
+          
         return programFile
       }
     )
@@ -160,20 +122,9 @@ export function activateDaffodilDebug(
   context.subscriptions.push(
     vscode.commands.registerCommand(
       'extension.dfdl-debug.getDataName',
-      async (config) => {
-        // If prgramFile is not set do not prompt for dataFile
-        const programFile = Buffer.from(
-          await vscode.workspace.fs.readFile(
-            vscode.Uri.parse(`${xdgAppPaths.data()}/.programFile`)
-          )
-        ).toString('utf8')
-        if (programFile === '') {
-          stopDebugging()
-          return ''
-        }
-
+      async (_) => {
         // Open native file explorer to allow user to select data file from anywhere on their machine
-        let dataFile = await vscode.window
+        return await vscode.window
           .showOpenDialog({
             canSelectMany: false,
             openLabel: 'Select input data file to debug',
@@ -188,13 +139,6 @@ export function activateDaffodilDebug(
 
             return ''
           })
-
-        // If data file not selected stop launch
-        if (dataFile === '') {
-          stopDebugging()
-        }
-
-        return dataFile
       }
     )
   )
